Updates to test
