+++ src/Services/JsonMapper.php
@@ -19,6 +19,7 @@ class JsonMapper
 
         $jsonDecodedData = json_decode($json, true);
         foreach ($jsonDecodedData as $key => $value) {
+
             if (!property_exists($newObject, $key)) {
                 continue;
             }
